Test opencascade/stl_refinement: do not compare output
The output of this test is a mesh where depending on the opencascade
version in question, toolchain and hardware a slightly different
interpolation is found. Let us not compare the final mesh and simply
rely on the routine to throw an error if it cannot create and refine the
mesh.