Rename a lambda function and make it easier to read.
The function is currently called 'sort_constraints()', but it also makes
the elements of a vector unique. So call it 'sort_and_make_unique()'
instead.

More importantly, though, the function operates on a vector that it captures
by reference. That makes it hard to read at the place where it is called what
it works on. Make this more obvious by letting it take the object it works on
as an argument.